+++ src/main/java/racingcar/RacingcarApplication.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.GameManager;
+
 public class RacingcarApplication {
     public static void main(String[] args) {
         GameManager gameManager = new GameManager();
